Use @QuarkusTest to be sure config is loaded
Add some basic tests for the QuarkusExtensionCatalog
